requirements.txt
argparse
paho-mqtt
pyyaml
requests
simplejson
simple-queue